[pf] Amend a008d11: Use more general term annotation in crits
Both Spezifikationen and Zusicherungen are Annotationen and as such
embedded in the program (code). In a008d11 I just included Zusicherungen
(Run-time-Checker) in the module description and did not check for
consistency with the detailed list of criteria below. To prevent this
from happening again just talk about Annotationen when specifying
Annotationssprache.

Every annotation type states something about the execution of an
annotated program. Whether its properness is checked for a single or any
execution is defined by whether it is checked by Run-time-Checker or
Beweiser. Read "mindestens einer Programmausführung".