[ph] Make Z3 model to program parameter translation Wunsch
As it is not yet clear whether a program specification verification can
always be done modular it might be difficult to trace back function
models to parameter assignments the program needs to be called with in
order to produce unexpected behaviour. Moreover in positive scenarios
the rest of the program must meet its specification, otherwise its
behaviour will be anything different from its conditions and maybe even
compensating the checked code's faulty behaviour.