[debugger,model,grammar] Add new language for expressione evaluation
As suggested in http://www.eclipse.org/forums/index.php/t/279319/, we
are now using a specialized language for expression parsing, which
avoids conflicts when using Google Guice.

ExpressionEvaluation now extends from ASTNode and not from Statement.

Also, the Worthwhile grammar is not cluttered with the __eval statement
any more.

Expression parsing is broken at the moment, but chances of fixing it
soon are much better than previously.