[prover,model] Add ArrayFunctionAccess
In the formula generated by WPStrategy we need to model both the value
of arrays and field accesses to these values. It is not possible to
replace field accesses with the respective value from the array literal
in visit(VariableReference) because index Expressions are not evaluated
when transformation is done.