[prover] Make substitute cloning optional
It is not always desirable to only return clones of what has been set by
SubstitutionVisitor::setSubstitute. For example this is not the case
when the substitute is instantiated for every replacement and would be
cloned immediately afterwards.

Originally the only SubstitutionVisitor implementation was for replacing
multiple VariableReferences with the same Expression substitute, which
was once set and therefore needed to be cloned for every new
replacement.