[prover] Only bind variables changed in loop for postcondition proof
Binding all variables that occur in the loop body is unneeded and
requires the specification of redundant information in invariants and
makes the formulas to prove loop postconditions very compilcated.

The binding only those variables that are actually changed with a forall
quantifier is legal because only for those variables we need to make the
admission that we have no idea what happened to them inside the loop.
For those that nothing has been assigned to, we know exactly what
happened - nothing!

This fixes #43.