[prover] Bind only function parameters instead of variables
Binding only the function parameters in the function precondition is
more correct because only for those can we really expect "any input".
There shouldn't be any other unbound variables left in the function
precondition, if there were, it would be a bug in our implementation and
binding variables that are only there because of a bug is wrong.