[prover] Fix #95: Insert variable "_return" for ReturnValueReferences
First search for a variable with the magic name "_return" in the
environment provided to checkFormula by a caller (the Worthwhile
interpreter at the moment) and if the environment does not specify one
create it. Second replace all ReturnValueReferences in the to be checked
formula with a VariableReference to the variable selected in the first
step.

This is implemented in SpecificationChecker since the prover cannot deal
with ReturnValueReferences because their semantics are not defined
unless given in the context of a FunctionDeclaration.
SpecificationChecker should be seen as the interpreter's interface to
the prover for generating first order logic formulae (only variables, no
FunctionCalls or ReturnValueReferences) from Annotations.