[prover] Fix FunCallSub ignores function ReturnValueReference refers to
FunctionCallSubstitution replaced all ReturnValueReferences in the
resolved postcondition regardless of which FunctionDeclaration (the
called or the calling one) they would refer to. This results in a error
when the calling function uses its ReturnValueReferences as actual
parameter in a FunctionCall and FunctionCallSubstitution replaces all
ReturnValueReferences with the return variable _after_ the actuals have
been inserted (the inserted ReturnValueReference is wrongly replaced
with the return variable).

Unless ReturnValueReferences have a reference to the FunctionDeclaration
they refer to (like VariableReferences have a reference to the
respective VariableDeclaration) we insert the references to the return
value before inserting the actual parameters. This is possible since the
calling function cannot refer to the called function's return value with
a ReturnValueReference.

- get(Precondition|Postcondition) from now on take the return value
  VariableDeclaration for the parameters mapping state instead of
  directly the FunctionCall they are mapped to, which makes the two
  methods FunctionCallSubstition-specific
- this fixes github issue #108