[model,grammar,ui] Fix #107: Reference function declaration in ReturnValueRef.
To achieve this, we create a new "fake" cross reference to the function
declaration in the grammar. This cross reference has the static name
"_return", which makes the language syntax stay the same.

Now we can use the WorthwhileScopeProvider to provide a scope that contains
a FunctionDeclaration with this name ("_return"). This function declaration
is now assigned to the reference.

Update the TokenToAttributeIdMapper since "_return" is now not automatically
recognized as a keyword any more.