[prover] Add macro substitution for simple postconditions
If the postcondition of a function is "simple", i.e. of the form
"_ensures _return = XYZ", we can replace a function call by the ensured
return value XYZ.

This avoids cluttered formulas when using helper functions.

The macro substitution is applied before the function call substitution
which substitutes all leftover function calls. However, the preconditions
need to be asserted in both cases. This functionality has been placed
in a special FunctionCallPreconditionAssertion class.